refactor: move event handler classes to event-handler directory
